python-service/requirements.txt
flask
scikit-learn
faiss-cpu
numpy
pandas
